+++ glide-core/redis-rs/redis/src/cluster_async/mod.rs
@@ -332,6 +332,12 @@ where
             .await
     }
 
+    /// Update the database ID used for all cluster connections
+    pub async fn update_connection_database(&mut self, database_id: i64) -> RedisResult<Value> {
+        self.route_operation_request(Operation::UpdateConnectionDatabase(database_id))
+            .await
+    }
+
     /// Update the name used for all cluster connections
     pub async fn update_connection_client_name(
         &mut self,
@@ -664,6 +670,7 @@ enum CmdArg<C> {
 #[derive(Clone)]
 enum Operation {
     UpdateConnectionPassword(Option<String>),
+    UpdateConnectionDatabase(i64),
     UpdateConnectionClientName(Option<String>),
     GetUsername,
 }
@@ -2599,6 +2606,11 @@ where
             CmdArg::OperationRequest(operation_request) => match operation_request {
                 Operation::UpdateConnectionPassword(password) => {
                     core.set_cluster_param(|params| params.password = password)
+                        .expect(MUTEX_WRITE_ERR);
+                    Ok(Response::Single(Value::Okay))
+                }
+                Operation::UpdateConnectionDatabase(database_id) => {
+                    core.set_cluster_param(|params| params.database_id = database_id)
                         .expect(MUTEX_WRITE_ERR);
                     Ok(Response::Single(Value::Okay))
                 }