+++ node/src/GlideClient.ts
@@ -16,7 +16,6 @@ import {
     createConfigResetStat,
     createConfigRewrite,
     createConfigSet,
-    createCopy,
     createCustomCommand,
     createDBSize,
     createEcho,
@@ -607,48 +606,6 @@ export class GlideClient extends BaseClient {
         return this.createWritePromise(createTime(), {
             decoder: Decoder.String,
         });
-    }
-
-    /**
-     * Copies the value stored at the `source` to the `destination` key. If `destinationDB` is specified,
-     * the value will be copied to the database specified, otherwise the current database will be used.
-     * When `replace` is true, removes the `destination` key first if it already exists, otherwise performs
-     * no action.
-     *
-     * @see {@link https://valkey.io/commands/copy/|valkey.io} for more details.
-     * @remarks Since Valkey version 6.2.0.
-     *
-     * @param source - The key to the source value.
-     * @param destination - The key where the value should be copied to.
-     * @param options - (Optional) Additional parameters:
-     * - (Optional) `destinationDB`: the alternative logical database index for the destination key.
-     *     If not provided, the current database will be used.
-     * - (Optional) `replace`: if `true`, the `destination` key should be removed before copying the
-     *     value to it. If not provided, no action will be performed if the key already exists.
-     * @returns `true` if `source` was copied, `false` if the `source` was not copied.
-     *
-     * @example
-     * ```typescript
-     * const result = await client.copy("set1", "set2");
-     * console.log(result); // Output: true - "set1" was copied to "set2".
-     * ```
-     * ```typescript
-     * const result = await client.copy("set1", "set2", { replace: true });
-     * console.log(result); // Output: true - "set1" was copied to "set2".
-     * ```
-     * ```typescript
-     * const result = await client.copy("set1", "set2", { destinationDB: 1, replace: false });
-     * console.log(result); // Output: true - "set1" was copied to "set2".
-     * ```
-     */
-    public async copy(
-        source: GlideString,
-        destination: GlideString,
-        options?: { destinationDB?: number; replace?: boolean },
-    ): Promise<boolean> {
-        return this.createWritePromise(
-            createCopy(source, destination, options),
-        );
     }
 
     /**